make updates for jest
